backend/ai-service/requirements.txt
crewai
langchain-groq
pika
python-dotenv
requests
chromadb
